requirements.txt
flask==3.0.0
requests==2.31.0
sqlite3
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
gunicorn==21.2.0
opencc-python-reimplemented==0.1.7
jieba==0.42.1 